events store
- magnific-popup isolated file
- handle packs depending admin or site
