+++ app/javascript/gobierto_TEMP/modules/dashboard_maker_controller.js
@@ -1,4 +1,5 @@
 import Maker from "../webapp/Maker.vue"
+// import "../../assets/stylesheets/module-TEMP-maker.scss";
 import { GobiertoDashboardCommonsController } from "./dashboard_commons_controller"
 
 export class GobiertoDashboardMakerController extends GobiertoDashboardCommonsController {
