+++ app/javascript/gobierto_TEMP/modules/dashboard_viewer_controller.js
@@ -1,4 +1,5 @@
 import Viewer from "../webapp/Viewer.vue"
+// import "../../assets/stylesheets/module-TEMP-viewer.scss";
 import { GobiertoDashboardCommonsController } from "./dashboard_commons_controller"
 
 export class GobiertoDashboardViewerController extends GobiertoDashboardCommonsController {
