+++ app/javascript/lib/commons/index.js
@@ -31,6 +31,3 @@ document.addEventListener("DOMContentLoaded", () => {
     Turbolinks.start()
   }
 })
-
-// TODO: Esto podría ser dividido en pequeños módulos para inyectar solo las cosas necesarias
-import 'lib/shared'
