+++ app/javascript/lib/shared/index.js
@@ -5,6 +5,7 @@ import "./modules/module-search.js";
 import "./modules/module-sessions.js";
 import "./modules/module-site_header.js";
 import "./modules/globals.js";
+import "./modules/magnific-popup.js";
 import d3locale from "./modules/d3-locale.js";
 import { isDesktop, isMobile } from "./modules/globals.js";
 import "./modules/tabs.js";
@@ -34,6 +35,7 @@ import {
   clickoutside
 } from "./modules/vue-directives.js";
 import { Middleware } from "./modules/middleware.js";
+import * as GobiertoEvents from "./modules/events.js";
 
 export {
   AUTOCOMPLETE_DEFAULTS,
@@ -55,5 +57,6 @@ export {
   truncate,
   percent,
   VueDirectivesMixin,
-  clickoutside
+  clickoutside,
+  GobiertoEvents
 };
