+++ app/javascript/lib/shared/modules/globals.js
@@ -1,4 +1,3 @@
-import "magnific-popup";
 import "tipsy-1a";
 
 // Global util functions
@@ -15,40 +14,6 @@ $.fn.extend({
   toggleText: function(a, b) {
     return this.text(this.text() == b ? a : b);
   }
-});
-
-$(document).on("turbolinks:load ajax:complete ajaxSuccess", function() {
-  $(".open_remote_modal").magnificPopup({
-    type: "ajax",
-    removalDelay: 300,
-    mainClass: "mfp-move-horizontal",
-    callbacks: {
-      ajaxContentAdded: function() {
-        // This conditionals are always true ¬¬
-        if (window.GobiertoAdmin && window.GobiertoAdmin.process_stages_controller) {
-          window.GobiertoAdmin.process_stages_controller.form();
-        }
-        if (window.GobiertoAdmin && window.GobiertoAdmin.globalized_forms_component) {
-          window.GobiertoAdmin.globalized_forms_component.handleGlobalizedForm();
-        }
-        if (window.GobiertoAdmin && window.GobiertoAdmin.gobierto_citizens_charters_editions_intervals_controller) {
-          window.GobiertoAdmin.gobierto_citizens_charters_editions_intervals_controller.handleForm();
-        }
-        if (window.GobiertoAdmin && window.GobiertoAdmin.admin_groups_admins_controller) {
-          window.GobiertoAdmin.admin_groups_admins_controller.index();
-        }
-        if (window.GobiertoAdmin && window.GobiertoAdmin.terms_controller) {
-          window.GobiertoAdmin.terms_controller.form()
-        }
-
-        // autofocus on the first modal input field
-        $(".modal .form_item input[type=text]:visible").first().focus()
-
-const event = new Event('create-dashboard-maker')
-document.dispatchEvent(event)
-      }
-    }
-  });
 });
 
 $(document).on("ajax:complete ajaxSuccess", function() {
@@ -78,17 +43,6 @@ $(document).on("turbolinks:load", function() {
       .addClass("active");
     scope.find("[data-tab]").removeClass("active");
     scope.find('[data-tab="' + target + '"]').addClass("active");
-  });
-
-  // Modal windows
-  $(".open_modal").magnificPopup({
-    type: "inline",
-    removalDelay: 300,
-    mainClass: "mfp-move-horizontal"
-  });
-
-  $(".close_modal").click(function() {
-    $.magnificPopup.close();
   });
 
   function toggleTarget($this) {
