+++ app/views/gobierto_admin/gobierto_temp/temp/index.html.erb
@@ -1,4 +1,3 @@
 <div class="modal modal-full-width modal-no-padding">
-  <h1>HELOa</h1>
   <div id="dashboard-maker-app" data-context="CONTEXTO" data-path="PATH"></div>
 </div>
