+++ app/views/sandbox/admin_TEMP.erb
@@ -1,5 +1,6 @@
 <% content_for :javascript_module_link do %>
-  <%= javascript_packs_with_chunks_tag 'TEMP' %>
+  <!-- No es necesario adjuntarlo xq ya está DENTRO del pack 'admin' -->
+  <!-- <%= javascript_packs_with_chunks_tag 'TEMP' %> -->
 <% end %>
 
 <% content_for :stylesheet_module_link do %>
@@ -8,8 +9,3 @@
 
 Lanzar modal:
 <a class="button open_remote_modal" href="/admin/gobierto_temp/index">Editar</a>
-
-<hr>
-
-Ejemplo del visor (usar el id <i>dashboard-viewer-app</i>)
-<div id="dashboard-viewer-app" data-context="CONTEXTO" data-path="PATH"></div>
